Refactored non-java stuff out of FormatterFactory.
