+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Java.java
@@ -15,58 +15,13 @@
  */
 package com.diffplug.spotless.maven.java;
 
-import static java.util.stream.Collectors.toList;
-
-import java.io.File;
-import java.nio.charset.Charset;
-import java.util.List;
-import java.util.Objects;
-
-import org.apache.maven.plugins.annotations.Parameter;
-
-import com.diffplug.spotless.FormatExceptionPolicyStrict;
-import com.diffplug.spotless.Formatter;
-import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.maven.FormatterFactory;
-import com.diffplug.spotless.maven.FormatterStepFactory;
-import com.diffplug.spotless.maven.MojoConfig;
 
-public class Java implements FormatterFactory {
-
+public class Java extends FormatterFactory {
 	private static final String EXTENSION = "java";
-
-	@Parameter
-	private String encoding;
-
-	@Parameter
-	private LineEnding lineEndings;
-
-	@Parameter
-	private List<FormatterStepFactory> steps;
 
 	@Override
 	public String fileExtension() {
 		return EXTENSION;
-	}
-
-	@Override
-	public Formatter newFormatter(List<File> filesToFormat, MojoConfig mojoConfig) {
-		Charset formatterEncoding = encoding == null ? Charset.forName(mojoConfig.getEncoding()) : Charset.forName(encoding);
-		LineEnding formatterLineEndings = lineEndings == null ? mojoConfig.getLineEndings() : lineEndings;
-		LineEnding.Policy formatterLineEndingPolicy = formatterLineEndings.createPolicy(mojoConfig.getBaseDir(), () -> filesToFormat);
-
-		List<FormatterStep> formatterSteps = steps.stream()
-				.filter(Objects::nonNull) // all unrecognized steps from XML config appear as nulls in the list
-				.map(factory -> factory.newFormatterStep(this, mojoConfig))
-				.collect(toList());
-
-		return Formatter.builder()
-				.encoding(formatterEncoding)
-				.lineEndingsPolicy(formatterLineEndingPolicy)
-				.exceptionPolicy(new FormatExceptionPolicyStrict())
-				.steps(formatterSteps)
-				.rootDir(mojoConfig.getBaseDir().toPath())
-				.build();
 	}
 }
